+++ src/onegov.feriennet/onegov/feriennet/path.py
@@ -2,15 +2,21 @@ from onegov.activity import Occasion, OccasionCollection
 from onegov.feriennet import FeriennetApp
 from onegov.feriennet.collections import VacationActivityCollection
 from onegov.feriennet.models import VacationActivity
+from onegov.feriennet.converters import age_range_converter
 from uuid import UUID
 
 
 @FeriennetApp.path(
     model=VacationActivityCollection,
     path='/angebote',
-    converters=dict(tags=[str], states=[str], durations=[int]))
+    converters=dict(
+        tags=[str],
+        states=[str],
+        durations=[int],
+        age_ranges=[age_range_converter]
+    ))
 def get_vacation_activities(request, app, page=0,
-                            tags=[], states=[], durations=[]):
+                            tags=[], states=[], durations=[], age_ranges=[]):
 
     return VacationActivityCollection(
         session=app.session(),
@@ -18,7 +24,8 @@ def get_vacation_activities(request, app, page=0,
         identity=request.identity,
         tags=tags,
         states=states,
-        durations=durations
+        durations=durations,
+        age_ranges=age_ranges
     )